refactor: dataend component is created and the order of the code in the components is optimized
