+++ src/components/CountriesList.js
@@ -1,45 +1,20 @@
-import React, { useContext, useState } from 'react';
-import { CountriesContext } from './context/CountriesContext'
+import React, { useContext } from 'react';
+import { CountriesContext } from './context/CountriesContext';
 import { Country } from './Country';
-import "../styles/CountryList.css"
+import { DataEnd } from './DataEnd';
 
 
 export const CountriesList = () => {
 
-  const {dataFilter} = useContext(CountriesContext)
-
-  const [inview, setInview] = useState(12);
-
-  let arrCountries = dataFilter?.slice(0, inview);
-
-  const showMoreItems = () => {
-    setInview((value) => value + 12);
-  };
-
-  const showLessItems = () => {
-    setInview((value) => value - 12);
-  };
+const {arrCountries} = useContext(CountriesContext);
 
 return(
 
-<div className='list'>
-    {arrCountries.length===0 ? <p>Por favor busca un pais </p> : arrCountries.map(countries=>(
-            <Country key={countries.code} {...countries}/> ))}
-
-    {dataFilter.length <= arrCountries.length ? (
-       null) : (
-      <div className='more'>   
-      <button onClick={showMoreItems} className="boton">
-        More
-      </button>
-      </div> 
-       )}
-
-    {arrCountries.length > 13 && (
-      <div className='less'>
-        <button onClick={showLessItems} className="boton">
-          Less
-        </button>
-      </div>)}
-</div>
-)}
+  <div className='list'>
+    {arrCountries.length===0 ? <p>no results were found with the name of the country entered </p> : arrCountries.map(countries=>(
+    
+    <Country key={countries.code} {...countries}/> ))};
+  
+    <DataEnd/>;
+  </div> 
+)};
