+++ src/components/NavBar.js
@@ -1,14 +1,13 @@
-import React from 'react'
-
-import "../styles/NavBar.css"
-import { SearchBar } from './Searchbar'
+import React from 'react';
+import "../styles/NavBar.css";
+import { SearchBar } from './Searchbar';
 
 export const NavBar = () => {
 
   return (
     <div className='fund'>
-     <h1 className='letter'>Country Search</h1>
-     <SearchBar/>
+      <h1 className='letter'>Country Search</h1>
+      <SearchBar/>
     </div>
   )
 }
