+++ src/components/Searchbar.js
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from 'react';
 import { CountriesContext } from './context/CountriesContext';
-import "../styles/SearchBar.css"
+import "../styles/SearchBar.css";
 
 
 export const SearchBar = () => {
@@ -16,9 +16,7 @@ export const SearchBar = () => {
  } 
   return (
     <form>
-      
        <input className="search" type="text" value={search} onChange={(e)=>handleChange(e)} placeholder="  Escribe el nombre de un pais..."></input>
-      
     </form>
   )
 }
