+++ src/styles/ContinentView.css
@@ -5,6 +5,6 @@ padding-left: 300px;
 }
 
 .continent h2{
+ margin-right:15%;   
  text-align: center;
- margin-right:15%   
 }
