+++ src/styles/CountryList.css
@@ -1,28 +1,82 @@
-.list{
- display:grid;
- grid-template-columns: 40% 40%;   
- padding-left: 300px;
-}
-
 .boton{
- width:200px;
- border-radius: 3px;
- padding:10px;
- margin-bottom: 10px;
- background-color: rgba(255, 200, 0, 0.516);  
+  background-color: rgba(255, 200, 0, 0.516);  
+  border-radius: 3px;
+  margin-bottom: 10px;
+  padding:10px;
+  width:200px;
 }
 
 .boton:hover{
-  cursor:pointer;
   background-color: rgba(9, 163, 184, 0.507);
-  color:black
+  color:black;
+  cursor:pointer;
+}
+.btn-end{
+  display:flex;
+  margin-left:40%;
 }
 
 .less{
   text-align:justify;  
 }
 
+.list{
+  display:grid;
+  grid-template-columns: 40% 40%;   
+  padding-left: 300px;
+}
+
 .more{
-text-align:right;
-margin-right: 90px;    
+  margin-right: 90px;    
+  text-align:right;
+}
+
+@media (max-width:425px){
+  .boton{
+    width:100px;
+  }
+  .list{
+    grid-template-columns: 50%;
+    justify-items: end;  
+  }
+}
+
+@media (min-width:425px) and (max-width:768px){
+  .boton{
+    margin-left: -120px; 
+    width:100px;
+  }
+  .less{
+    margin-left: 40px;  
+  }
+  .list{
+    grid-template-columns: 50%;   
+    padding-left: 150px;
+  }
+  .more{
+    margin-right: 90px;    
+  }
+}
+
+@media (min-width:768px) and (max-width:1024px){
+  .boton{
+    margin-left: -90px; 
+    width:200px;
+  }
+  .list{
+    grid-template-columns: 40% 40%; 
+    justify-content: space-between;
+  }
+}
+    
+@media (min-width:1024px) and (max-width:1439px){
+  .boton{
+    margin-left: 0px; 
+    width:200px;
+  }
+  .list{
+    grid-template-columns: 50% 50%;
+    justify-content: space-between;
+    margin-left: -200px; 
+  }
 }
