+++ src/styles/Filter.css
@@ -1,31 +1,46 @@
 .filter{
- height:6rem; 
- display: flex;
  align-items: center;
+ display: flex;
+ height:6rem; 
  justify-content: center;  
 }
 
 .filter button{
-  margin-left: 20px;
-  text-decoration:none;
-  font-size: 20px;
   border-radius: 3px;
-  text-align: center;
-  width:200px;
+  font-size: 20px;
   font-weight: bold;
+  margin-left: 20px;
+  text-align: center;
+  text-decoration:none;
+  width:200px;
 }
 
 .filter button:hover{
-  cursor:pointer;
   background-color: #0a0a3c;
   color:white; 
+  cursor:pointer;
 }
 
 .filter label{
-    font-size: 50px;
+  font-size: 50px;
 }
 
 .filter button Link{
   text-decoration: none;
   text-align: center;  
+}
+
+
+@media (max-width:768px){
+.filter label{
+  font-size: 20px;
+}
+.filter button{
+  border-radius: 3px;
+  font-size: 15px;
+  font-weight: bold;
+  text-align: center;
+  text-decoration:none;
+  width:100px;
+}
 }
