+++ src/styles/NavBar.css
@@ -6,14 +6,26 @@ body{
   font-family: 'Raleway', sans-serif; 
 }
 .fund{
+  align-items: center;
   background-color: #0a0a3c;
   display:flex;
   height:5rem; 
-  align-items: center;
   justify-content: space-evenly;
 }
 
 .letter{
-    color:white;
-    font-size: 50px;
+  color:white;
+  font-size: 50px;
+}
+
+@media (max-width:768px){
+.letter{
+  font-size: 20px;
+  }
+
+.fund{
+  margin-right: 0; 
+  width:100vw;
+  }
+
 }
