+++ src/styles/SearchBar.css
@@ -1,5 +1,11 @@
 .search{
-  width:30rem;
-  height:2.5rem;
   border-radius: 5px;  
+  height:2.5rem;
+  width:30rem;
+}
+
+@media (max-width:768px){
+  .search{
+    width:16rem;
+  }
 }
